revise print for ordinal models

+++ R/print.R
@@ -32,6 +32,7 @@ print.ggeffects <- function(x, n = 10, digits = 3, x.lab = FALSE, ...) {
   has_groups <- obj_has_name(x, "group") && length(unique(x$group)) > 1
   has_facets <- obj_has_name(x, "facet") && length(unique(x$facet)) > 1
   has_panel <- obj_has_name(x, "panel") && length(unique(x$panel)) > 1
+  has_response <- obj_has_name(x, "response.level") && length(unique(x$response.level)) > 1
   has_se <- obj_has_name(x, "std.error")
 
   cat("\n")
@@ -84,6 +85,11 @@ print.ggeffects <- function(x, n = 10, digits = 3, x.lab = FALSE, ...) {
     }
   }
 
+  if (has_response) {
+    .n <- .n * .n_distinct(x$response.level)
+    x$response.level <- sprintf("Response Level = %s", as.character(x$response.level))
+  }
+
   # make sure that by default not too many rows are printed
   if (missing(n)) {
     n <- dplyr::case_when(
@@ -95,36 +101,79 @@ print.ggeffects <- function(x, n = 10, digits = 3, x.lab = FALSE, ...) {
   }
 
   if (!has_groups) {
-    cat("\n")
-    if (obj_has_name(x, "group"))
-      x <- .remove_column(x, "group")
-    print.data.frame(x[get_sample_rows(x, n), ], ..., row.names = FALSE, quote = FALSE)
+
+    if (!has_response) {
+      cat("\n")
+      if (obj_has_name(x, "group")) x <- .remove_column(x, "group")
+      print.data.frame(x[.get_sample_rows(x, n), ], ..., row.names = FALSE, quote = FALSE)
+    } else {
+      x$.nest <- tapply(x$predicted, list(x$response.level), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n", i$response.level[1]), "red")
+        .print_block(i, n, ...)
+      }
+    }
+
   } else if (has_groups && !has_facets) {
-    x$.nest <- tapply(x$predicted, list(x$group), NULL)
-    xx <- split(x, x$.nest)
 
-    for (i in xx) {
-      insight::print_color(sprintf("\n# %s\n", i$group[1]), "red")
-      i <- i[setdiff(colnames(i), c("group", "facet", "panel", ".nest"))]
-      print.data.frame(i[get_sample_rows(i, n), ], ..., row.names = FALSE, quote = FALSE)
+    if (!has_response) {
+      x$.nest <- tapply(x$predicted, list(x$group), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n", i$group[1]), "red")
+        .print_block(i, n, ...)
+      }
+    } else {
+      x$.nest <- tapply(x$predicted, list(x$response.level, x$group), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n# %s\n", i$response.level[1], i$group[1]), "red")
+        .print_block(i, n, ...)
+      }
     }
+
   } else if (has_groups && has_facets && !has_panel) {
-    x$.nest <- tapply(x$predicted, list(x$group, x$facet), NULL)
-    xx <- split(x, x$.nest)
 
-    for (i in xx) {
-      insight::print_color(sprintf("\n# %s\n# %s\n", i$group[1], i$facet[1]), "red")
-      i <- i[setdiff(colnames(i), c("group", "facet", "panel", ".nest"))]
-      print.data.frame(i[get_sample_rows(i, n), ], ..., row.names = FALSE, quote = FALSE)
+    if (!has_response) {
+      x$.nest <- tapply(x$predicted, list(x$group, x$facet), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n# %s\n", i$group[1], i$facet[1]), "red")
+        .print_block(i, n, ...)
+      }
+    } else {
+      x$.nest <- tapply(x$predicted, list(x$response.level, x$group, x$facet), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n# %s\n# %s\n", i$response.level[1], i$group[1], i$facet[1]), "red")
+        .print_block(i, n, ...)
+      }
     }
-  } else {
-    x$.nest <- tapply(x$predicted, list(x$group, x$facet, x$panel), NULL)
-    xx <- split(x, x$.nest)
 
-    for (i in xx) {
-      insight::print_color(sprintf("\n# %s\n# %s\n# %s\n", i$group[1], i$facet[1], i$panel[1]), "red")
-      i <- i[setdiff(colnames(i), c("group", "facet", "panel", ".nest"))]
-      print.data.frame(i[get_sample_rows(i, n), ], ..., row.names = FALSE, quote = FALSE)
+  } else {
+
+    if (!has_response) {
+      x$.nest <- tapply(x$predicted, list(x$group, x$facet, x$panel), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n# %s\n# %s\n", i$group[1], i$facet[1], i$panel[1]), "red")
+        .print_block(i, n, ...)
+      }
+    } else {
+      x$.nest <- tapply(x$predicted, list(x$response.level, x$group, x$facet, x$panel), NULL)
+      xx <- split(x, x$.nest)
+
+      for (i in xx) {
+        insight::print_color(sprintf("\n# %s\n# %s\n# %s\n# %s\n", i$response.level[1], i$group[1], i$facet[1], i$panel[1]), "red")
+        .print_block(i, n, ...)
+      }
     }
   }
 
@@ -176,7 +225,8 @@ print.ggeffects <- function(x, n = 10, digits = 3, x.lab = FALSE, ...) {
 }
 
 
-get_sample_rows <- function(x, n) {
+
+.get_sample_rows <- function(x, n) {
   nr.of.rows <- sjmisc::seq_row(x)
 
   if (n < length(nr.of.rows)) {
@@ -190,4 +240,11 @@ get_sample_rows <- function(x, n) {
   }
 
   sample.rows
+}
+
+
+
+.print_block <- function(i, n, ...) {
+  i <- i[setdiff(colnames(i), c("group", "facet", "panel", "response.level", ".nest"))]
+  print.data.frame(i[.get_sample_rows(i, n), ], ..., row.names = FALSE, quote = FALSE)
 }
